refactor(session-notification): move cleanupOldSessions to utils and add edge case tests
Move cleanupOldSessions from platform module to utils module for
better separation of concerns. Add tests for empty format string
and missing template variables edge cases.
